+++ cpp/include/cuspatial/spatial_join.hpp
@@ -25,8 +25,6 @@ namespace cuspatial {
 /**
  * @brief Search a quadtree for polygon or polyline bounding box intersections.
  *
- * @note Swaps `x_min` and `x_max` if `x_min > x_max`.
- * @note Swaps `y_min` and `y_max` if `y_min > y_max`.
  * @note `scale` is applied to (x - x_min) and (y - y_min) to convert coordinates into a Morton code
  * in 2D space.
  * @note `max_depth` should be less than 16, since Morton codes are represented as `uint32_t`.
@@ -39,6 +37,7 @@ namespace cuspatial {
  * @param y_max The upper-right y-coordinate of the area of interest bounding box.
  * @param scale Scale to apply to each x and y distance from x_min and y_min.
  * @param max_depth Maximum quadtree depth at which to stop testing for intersections.
+ * @param mr The optional resource to use for output device memory allocations.
  *
  * @throw cuspatial::logic_error If the quadtree table is malformed
  * @throw cuspatial::logic_error If the polygon bounding box table is malformed
@@ -51,44 +50,111 @@ namespace cuspatial {
  * poly_offset - INT32 column of indices for each poly bbox that intersects with the quadtree.
  * quad_offset - INT32 column of indices for each leaf quadrant intersecting with a poly bbox.
  */
-
 std::unique_ptr<cudf::table> quad_bbox_join(
-  cudf::table_view const &quadtree,
-  cudf::table_view const &poly_bbox,
+  cudf::table_view const& quadtree,
+  cudf::table_view const& poly_bbox,
   double x_min,
   double x_max,
   double y_min,
   double y_max,
   double scale,
   int8_t max_depth,
-  rmm::mr::device_memory_resource *mr = rmm::mr::get_default_resource());
+  rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource());
 
 /**
- * @brief pair points and polygons using ray-cast based point-in-polygon test algorithm in two
- *phases: phase 1 counts the total number of output pairs for precise memory allocation phase 2
- *actually writes (point,polygon) pairs
+ * @brief Test whether the specified points are inside any of the specified polygons.
  *
- * @param pq_pair: table of two arrays for (quadrant,polygon) pairs derived from spatial filtering
- * @param quadtree: table of five arrays derived from quadtree indexing on points (key,lev,sign,
- * length, fpos)
- * @param pnt: table of two arrays for points (x,y). note that points are in-place sorted in
- * quadtree construction and have different orders than the orginal input points.
- * @param fpos: feature/polygon offset array to rings
- * @param rpos: ring offset array to vertex
- * @param poly_x: polygon x coordiante array.
- * @param poly_y: polygon y coordiante array.
+ * Uses the table of (polygon, quadrant) pairs returned by `cuspatial::quad_bbox_join` to
+ * ensure only the points in the same quadrant as each polygon are tested for intersection.
  *
- * @return array of (polygon_index, point_index) pairs that point is within polyon;
- * point_index and polygon_index are offsets of point and polygon arrays, respectively
- */
-// std::unique_ptr<cudf::table> pip_refine(
-//   cudf::table_view const& pq_pair,
-//   cudf::table_view const& quadtree,
-//   cudf::table_view const& pnt,
-//   cudf::column_view const& poly_fpos,
-//   cudf::column_view const& poly_rpos,
-//   cudf::column_view const& poly_x,
-//   cudf::column_view const& poly_y,
-//   rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource());
+ * This pre-filtering can dramatically reduce number of points tested per polygon, enabling
+ * faster intersection-testing at the expense of extra memory allocated to store the quadtree and
+ * sorted point_indices.
+ *
+ * @param poly_quad_pairs cudf table of (polygon, quadrant) index pairs returned by
+ * `cuspatial::quad_bbox_join`
+ * @param quadtree cudf table representing a quadtree (key, level, is_quad, length, offset).
+ * @param point_indices Sorted point indices returned by `cuspatial::quadtree_on_points`
+ * @param point_x x-coordinates of points to test
+ * @param point_y y-coordinates of points to test
+ * @param poly_offsets Begin indices of the first ring in each polygon (i.e. prefix-sum).
+ * @param ring_offsets Begin indices of the first point in each ring (i.e. prefix-sum).
+ * @param poly_points_x Polygon point x-coodinates.
+ * @param poly_points_y Polygon point y-coodinates.
+ * @param mr The optional resource to use for output device memory allocations.
+ *
+ * @throw cuspatial::logic_error If the poly_quad_pairs table is malformed.
+ * @throw cuspatial::logic_error If the quadtree table is malformed.
+ * @throw cuspatial::logic_error If the number of point indices doesn't match the number of points.
+ * @throw cuspatial::logic_error If the number of rings is less than the number of polygons.
+ * @throw cuspatial::logic_error If any ring has fewer than three vertices.
+ * @throw cuspatial::logic_error If the types of point and polygon vertices are different.
+ *
+ * @return A cudf table with two columns, where each row represents a point/polygon intersection:
+ *   point_offset - UINT32 column of point indices
+ * polygon_offset - UINT32 column of polygon indices
+ *
+ * @note The returned point and polygon indices are offsets into the (point_x, point_y) and
+ * poly_quad_pairs inputs, respectively.
+ *
+ **/
+std::unique_ptr<cudf::table> quadtree_point_in_polygon(
+  cudf::table_view const& poly_quad_pairs,
+  cudf::table_view const& quadtree,
+  cudf::column_view const& point_indices,
+  cudf::column_view const& point_x,
+  cudf::column_view const& point_y,
+  cudf::column_view const& poly_offsets,
+  cudf::column_view const& ring_offsets,
+  cudf::column_view const& poly_points_x,
+  cudf::column_view const& poly_points_y,
+  rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource());
+
+/**
+ * @brief Finds the nearest polyline to each point in a quadrant, and computes the distances between
+ * each point and polyline.
+ *
+ * Uses the table of (polyline, quadrant) pairs returned by `cuspatial::quad_bbox_join` to
+ * ensure distances are computed only for the points in the same quadrant as each polyline.
+ *
+ * @param poly_quad_pairs cudf table of (polyline, quadrant) index pairs returned by
+ * `cuspatial::quad_bbox_join`
+ * @param quadtree cudf table representing a quadtree (key, level, is_quad, length, offset).
+ * @param point_indices Sorted point indices returned by `cuspatial::quadtree_on_points`
+ * @param point_x x-coordinates of points to test
+ * @param point_y y-coordinates of points to test
+ * @param poly_offsets Begin indices of the first point in each polyline (i.e. prefix-sum)
+ * @param poly_points_x Polyline point x-coordinates
+ * @param poly_points_y Polyline point y-coordinates
+ * @param mr The optional resource to use for output device memory allocations.
+ *
+ * @throw cuspatial::logic_error If the poly_quad_pairs table is malformed.
+ * @throw cuspatial::logic_error If the quadtree table is malformed.
+ * @throw cuspatial::logic_error If the number of point indices doesn't match the number of points.
+ * @throw cuspatial::logic_error If any polyline has fewer than two vertices.
+ * @throw cuspatial::logic_error If the types of point and polyline vertices are different.
+ *
+ * @return A cudf table with three columns, where each row represents a point/polyline pair and the
+ * distance between the two:
+ *
+ *    point_offset - UINT32 column of point indices
+ * polyline_offset - UINT32 column of polyline indices
+ *        distance - FLOAT or DOUBLE column (based on input point data type) of distances between
+ *                   each point and polyline
+ *
+ * @note The returned point and polyline indices are offsets into the (point_x, point_y) and
+ * poly_quad_pairs inputs, respectively.
+ *
+ **/
+std::unique_ptr<cudf::table> quadtree_point_to_nearest_polyline(
+  cudf::table_view const& poly_quad_pairs,
+  cudf::table_view const& quadtree,
+  cudf::column_view const& point_indices,
+  cudf::column_view const& point_x,
+  cudf::column_view const& point_y,
+  cudf::column_view const& poly_offsets,
+  cudf::column_view const& poly_points_x,
+  cudf::column_view const& poly_points_y,
+  rmm::mr::device_memory_resource* mr = rmm::mr::get_default_resource());
 
 }  // namespace cuspatial