+++ python/cuspatial/cuspatial/core/spatial_join.py
@@ -3,8 +3,11 @@
 import warnings
 
 from cudf import DataFrame
+from cudf.core.column import as_column
 
 from cuspatial._lib import spatial_join
+
+from cuspatial.utils.column_utils import normalize_point_columns
 
 
 def quad_bbox_join(
@@ -37,9 +40,9 @@ def quad_bbox_join(
         Indices for each intersecting bounding box and leaf quadrant.
 
         poly_offset : cudf.Series
-            Indices for each poly bbox that intersects with the quadtree
+            Indices for each poly bbox that intersects with the quadtree.
         quad_offset : cudf.Series
-            Indices for each leaf quadrant intersecting with a poly bbox
+            Indices for each leaf quadrant intersecting with a poly bbox.
 
     Notes
     -----
@@ -70,5 +73,161 @@ def quad_bbox_join(
             y_max,
             max(scale, min_scale),
             max_depth,
+        )
+    )
+
+
+def quadtree_point_in_polygon(
+    poly_quad_pairs,
+    quadtree,
+    point_indices,
+    points_x,
+    points_y,
+    poly_offsets,
+    ring_offsets,
+    poly_points_x,
+    poly_points_y,
+):
+    """ Test whether the specified points are inside any of the specified
+    polygons.
+
+    Uses the table of (polygon, quadrant) pairs returned by
+    ``cuspatial.quad_bbox_join`` to ensure only the points in the same
+    quadrant as each polygon are tested for intersection.
+
+    This pre-filtering can dramatically reduce number of points tested per
+    polygon, enabling faster intersection-testing at the expense of extra
+    memory allocated to store the quadtree and sorted point_indices.
+
+    Parameters
+    ----------
+    poly_quad_pairs: cudf.DataFrame
+        Table of (polygon, quadrant) index pairs returned by
+        ``cuspatial.quad_bbox_join``.
+    quadtree : cudf.DataFrame
+        A complete quadtree for a given area-of-interest bounding box.
+    point_indices : cudf.Series
+        Sorted point indices returned by ``cuspatial.quadtree_on_points``
+    points_x : cudf.Series
+        x-coordinates of points used to construct the quadtree.
+    points_y : cudf.Series
+        y-coordinates of points used to construct the quadtree.
+    poly_offsets : cudf.Series
+        Begin index of the first ring in each polygon.
+    ring_offsets : cudf.Series
+        Begin index of the first point in each ring.
+    poly_points_x : cudf.Series
+        Polygon point x-coodinates.
+    poly_points_y : cudf.Series
+        Polygon point y-coodinates.
+
+    Returns
+    -------
+    result : cudf.DataFrame
+        Indices for each intersecting point and polygon pair.
+
+        point_offset : cudf.Series
+            Indices of each point that intersects with a polygon.
+        polygon_offset : cudf.Series
+            Indices of each polygon with which a point intersected.
+    """
+
+    (
+        points_x,
+        points_y,
+        poly_points_x,
+        poly_points_y,
+    ) = normalize_point_columns(
+        as_column(points_x),
+        as_column(points_y),
+        as_column(poly_points_x),
+        as_column(poly_points_y),
+    )
+    return DataFrame._from_table(
+        spatial_join.quadtree_point_in_polygon(
+            poly_quad_pairs,
+            quadtree,
+            as_column(point_indices, dtype="uint32"),
+            points_x,
+            points_y,
+            as_column(poly_offsets, dtype="uint32"),
+            as_column(ring_offsets, dtype="uint32"),
+            poly_points_x,
+            poly_points_y,
+        )
+    )
+
+
+def quadtree_point_to_nearest_polyline(
+    poly_quad_pairs,
+    quadtree,
+    point_indices,
+    points_x,
+    points_y,
+    poly_offsets,
+    poly_points_x,
+    poly_points_y,
+):
+    """ Finds the nearest polyline to each point in a quadrant, and computes
+    the distances between each point and polyline.
+
+    Uses the table of (polyline, quadrant) pairs returned by
+    ``cuspatial.quad_bbox_join`` to ensure distances are computed only for
+    the points in the same quadrant as each polyline.
+
+    Parameters
+    ----------
+    poly_quad_pairs: cudf.DataFrame
+        Table of (polyline, quadrant) index pairs returned by
+        ``cuspatial.quad_bbox_join``.
+    quadtree : cudf.DataFrame
+        A complete quadtree for a given area-of-interest bounding box.
+    point_indices : cudf.Series
+        Sorted point indices returned by ``cuspatial.quadtree_on_points``
+    points_x : cudf.Series
+        x-coordinates of points used to construct the quadtree.
+    points_y : cudf.Series
+        y-coordinates of points used to construct the quadtree.
+    poly_offsets : cudf.Series
+        Begin index of the first point in each polyline.
+    poly_points_x : cudf.Series
+        Polyline point x-coodinates.
+    poly_points_y : cudf.Series
+        Polyline point y-coodinates.
+
+    Returns
+    -------
+    result : cudf.DataFrame
+        Indices for each point and its nearest polyline, and the distance
+        between the two.
+
+        point_offset : cudf.Series
+            Indices of each point that intersects with a polyline.
+        polyline_offset : cudf.Series
+            Indices of each polyline with which a point intersected.
+        distance : cudf.Series
+            Distances between each point and its nearest polyline.
+    """
+    (
+        points_x,
+        points_y,
+        poly_points_x,
+        poly_points_y,
+    ) = normalize_point_columns(
+        as_column(points_x),
+        as_column(points_y),
+        as_column(poly_points_x),
+        as_column(poly_points_y),
+    )
+    return DataFrame._from_table(
+        spatial_join.quadtree_point_to_nearest_polyline(
+            poly_quad_pairs,
+            quadtree,
+            as_column(point_indices, dtype="uint32"),
+            points_x,
+            points_y,
+            as_column(poly_offsets, dtype="uint32"),
+            poly_points_x,
+            poly_points_y,
         )
     )